backend/requirements.txt
# Core
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-multipart>=0.0.20

# Database
sqlalchemy>=2.0.0

# Auth / Security
PyJWT>=2.10.1  # 2.10.0 has GHSA-75c5-xw7c-p5pm (iss claim partial match)
bcrypt>=4.2.0  # Using directly; passlib 1.7.4 is incompatible with bcrypt 4+
# Pin filelock to avoid GHSA-qmgc-5h2g-mvrw and GHSA-w853-jp5j-5j7f (symlink TOCTOU)
filelock>=3.20.3

# Settings
pydantic-settings>=2.8.0

# Rate limiting
slowapi>=0.1.9

# Audio processing / speech-to-text
faster-whisper>=1.1.0
mutagen>=1.47.0
filetype>=1.2.0

# HTML sanitization
nh3>=0.2.0

# HTTP client (for Ollama)
httpx>=0.28.0

# Testing
pytest>=8.3.0
pytest-asyncio>=0.25.0
pytest-cov>=6.0.0
anyio>=4.9.0

# Code quality
ruff>=0.9.0
pre-commit>=4.0.0
